+++ webpack.prod.js
@@ -26,7 +26,8 @@ module.exports = merge(common, {
         MODE: `"${MODE}"`,
         ENDPOINT_URL: `"https://xpqeqfjgwd.execute-api.us-east-1.amazonaws.com/v2/"`,
         USER_POOL_ID: `"us-east-1_kcpcLxLzn"`,
-        COGNITO_CLIENT_ID: `"77mcm1k9ll2ge68806h5kncfus"`
+        COGNITO_CLIENT_ID: `"77mcm1k9ll2ge68806h5kncfus"`,
+        GA_TRACKING_ID: `"UA-28518772-9"`
     }),
     new HtmlWebpackPlugin({
       title: 'Chinmaya Forms Framework',